🚚 [Move]: #154 numeric-operand.ts を numeric-pdf-object.ts にリネーム
ファイル名を export している型 NumericPdfObject に揃える。
matrix.ts / graphics-state.ts / line-cap.ts など既存ファイルの
「型名 = ファイル名 (kebab-case)」規則と一貫性を取る。
