+++ packages/core/src/content-stream/operators/graphics-state/numeric-pdf-object.ts
